Change GitHub source links to prevent link rot
Previously the commit reference was generated with rev-parse --abbrev-ref. This would return a name that wasn't actually fixed to the proper commit. For instance, when run at the most recent commit, it returns the branch name (i.e., master). Links using that will always point to the most recent commit in that branch, even though the source files change and lines move.

This makes links not point to where they're supposed to go, and can cause them to fail entirely if the file is renamed.

Using the --short option returns the first 7 characters of the commit SHA1 hash, which GitHub will resolve to be the exact point that the documentation was generated at, even after more commits are made later.